+++ examples/resources/iosxr_evpn/import.sh
@@ -1,0 +1,1 @@
+terraform import iosxr_evpn.example "Cisco-IOS-XR-um-l2vpn-cfg:/evpn"