+++ examples/resources/iosxr_evpn/resource.tf
@@ -1,0 +1,3 @@
+resource "iosxr_evpn" "example" {
+  source_interface = "Loopback0"
+}