+++ docs/guides/changelog.md
@@ -24,6 +24,7 @@ description: |-
 - Add `iosxr_evpn_evi` resource and data source
 - Add `iosxr_evpn_group` resource and data source
 - Add `iosxr_evpn_interface` resource and data source
+- Add `iosxr_evpn` resource and data source
 
 ## 0.1.4